Experimenting with PsiElementEvaluator [skip ci]
